docs: add docs and rearrange test
