+++ .gitignore
@@ -31,6 +31,7 @@ jspm_packages/
 *.tgz
 .yarn-integrity
 .cache
+.DS_Store
 !/.projenrc.js
 /test-reports/
 junit.xml
