+++ test/dummy.test.ts
@@ -3,3 +3,133 @@ import "jest";
 test("dummy", () => {
   expect(1).toBe(1);
 });
+
+import { DynamoDB } from "aws-sdk";
+
+import { TypeSafeDynamoDB, ApplyProjection } from "../src";
+
+export interface MyItem {
+  pk: string;
+  sk: string;
+  attr: string;
+  list?: MyItem[];
+}
+
+const table = new DynamoDB() as unknown as TypeSafeDynamoDB<MyItem, "pk", "sk">;
+
+export async function foo() {
+  const response = await table
+    .query({
+      TableName: "TableName",
+      AttributesToGet: ["attr"],
+      KeyConditionExpression: "#key = :val",
+      ProjectionExpression: "#key2 = :val2",
+      ExpressionAttributeNames: {
+        "#key": "",
+        "#key2": "",
+      },
+      ExpressionAttributeValues: {
+        ":val": {
+          S: "val",
+        },
+        ":val2": {
+          BOOL: true,
+        },
+      },
+    })
+    .promise();
+
+  if (response.Items) {
+    response.Items[0].attr;
+  }
+
+  await table
+    .getItem({
+      TableName: "",
+      Key: {
+        pk: {
+          S: "",
+        },
+        sk: {
+          S: "",
+        },
+      },
+      ProjectionExpression: "attr",
+    })
+    .promise();
+
+  await table
+    .putItem({
+      TableName: "",
+      Item: {
+        pk: {
+          S: "pk",
+        },
+        sk: {
+          S: "sk",
+        },
+        attr: {
+          S: "attr",
+        },
+        list: {
+          L: [
+            {
+              M: {
+                pk: {
+                  S: "pk",
+                },
+                sk: {
+                  S: "sk",
+                },
+                attr: {
+                  S: "attr",
+                },
+              },
+            },
+          ],
+        },
+      },
+      ConditionExpression: "#name = :value and #n = 0",
+      ExpressionAttributeValues: {
+        ":value": {
+          N: "1",
+        },
+      },
+      ExpressionAttributeNames: {
+        "#name": "sam",
+        "#n": "sam",
+      },
+    })
+    .promise();
+
+  interface Record {
+    id: string;
+    record?: Record;
+    items: (Record | undefined)[];
+    tuple: [a: "string", b: number];
+  }
+
+  type A1 = ApplyProjection<Record, "id">;
+  const a1: A1 = null as any;
+  a1.id;
+  type A2 = ApplyProjection<Record, "record">;
+  const a2: A2 = null as any;
+  a2.record?.id;
+  type A3 = ApplyProjection<Record, "record.id">;
+  const a3: A3 = null as any;
+  a3.record?.id;
+
+  type A4 = ApplyProjection<Record, "items[2], items[3]">;
+  const a4: A4 = null as any;
+  a4.items[0];
+
+  type A5 = ApplyProjection<Record, "tuple[1], tuple[0]">;
+  const a5: A5 = null as any;
+  a5.tuple[0];
+  a5.tuple[1];
+
+  type A6 = ApplyProjection<Record, "tuple[0], tuple[1]">;
+  const a6: A6 = null as any;
+  a6.tuple[0];
+  a6.tuple[1];
+}
